Rename rdkit.py in resonance to avoid confusing imports of rdkit
