+++ rdmc/resonance/__init__.py
@@ -3,5 +3,5 @@
 
 from rdmc.resonance.base import generate_resonance_structures
 # Make sure resonance algorithms are registered
-import rdmc.resonance.rdkit
-import rdmc.resonance.rmg
+import rdmc.resonance.rdkit_backend
+import rdmc.resonance.rmg_backend
