+++ test/resonance/test_rmg_backend.py
@@ -1,6 +1,6 @@
 from rdmc import RDKitMol
 from rdmc.resonance.utils import get_lone_pair, is_aromatic, is_identical
-from rdmc.resonance.rmg import (
+from rdmc.resonance.rmg_backend import (
     _clar_optimization,
     _clar_transformation,
     generate_clar_structures,
